added separate files sortable parents and child layers, and refactored code for sortLayerOrder
